pre-release for icingaweb_database_update module
update molecule tests
