+++ molecule/update_2.7.0-2.11.3/converge.yml
@@ -1,21 +1,21 @@
 ---
 
-- name: converge
+- name: converge for version 2.7.0
   hosts: instance
   any_errors_fatal: false
 
   vars:
-    icingaweb_version: 2.8.2
+    icingaweb_version: 2.7.0
 
   roles:
     - role: ansible-icingaweb2
 
-- name: update icingaweb
+- name: update icingaweb to 2.11.3
   hosts: instance
   any_errors_fatal: false
 
   vars:
-    icingaweb_version: 2.9.3
+    icingaweb_version: 2.11.3
 
   roles:
     - role: ansible-icingaweb2
