+++ molecule/update_2.7.0-2.11.3/molecule.yml
@@ -7,7 +7,7 @@ driver:
 
 platforms:
   - name: database
-    image: "bodsch/ansible-${DISTRIBUTION:-debian:10}"
+    image: "bodsch/ansible-debian:11"
     command: ${MOLECULE_DOCKER_COMMAND:-""}
     docker_host: "${DOCKER_HOST:-unix://run/docker.sock}"
     privileged: true
@@ -26,7 +26,7 @@ platforms:
         ipv4_address: "10.11.0.1"
 
   # - name: icinga2
-  #   image: "bodsch/ansible-${DISTRIBUTION:-debian:10}"
+  #   image: "bodsch/ansible-debian:11"
   #   command: ${MOLECULE_DOCKER_COMMAND:-""}
   #   docker_host: "${DOCKER_HOST:-unix://run/docker.sock}"
   #   privileged: true
